+++ CHANGELOG.md
@@ -9,9 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- moved `contains_any_types` over from mex-backend
+
 ### Changes
 
 - ensure extracted items are merged in predictable way
+- improve a batch of doc-strings with args, raises and return sections
 
 ### Deprecated