+++ CHANGELOG.md
@@ -22,6 +22,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- fix ldap error if ldap connector is >1h old
+
 ### Security
 
 ## [0.62.0] - 2025-06-17